Update with half-working shuffling
There are many many hacks going on here, some of them
are probably unnecessary, but it's hard to keep track.
The primary thing to work around is udon's vrchat's awfully
inconsistent networked physics sync on non-kinematic rigidbodies.
However that's implemented, it just doesn't work reliably.

So right now there are some manual buttons to press for the master,
and they'll have to rely on voice to confirm that the tiles aren't
fucked up for all the other players.