Add new syncing strategy with test
AFter a bunch of uploads and staring at logs, I think it works reliably enough
for multiple tables. Definitely easier to generalize on than the old system,
though it'll be kind of a mess for late joiners. I think the states should
settle in a few seconds. Very hard to test that assumption however.

and update VRCSDK. I think I might've changed some other things too
in the mmain world as well. git is too hard.