Improve tile UVs, make tile back color configurable
both at the material and the material property block, so I
can use the tile back as a visual debugging tool.

Removed the little chunk out of the sides; the low resolution of
the tile texture is hard to make it look nice. Though of course
I'm staring really closely at it, for regular use it doesn't matter.