Rewrite ownership system using lamport clocks
I was being a little stupid the first time with the
server timestamps. This way is a lot more robust, and
got rid of some of the other weirdness and kludges like
sending the shuffle packet over and over again.

Furthermore, fixed a large but unnoticed race condition with
the locallyMovedTiles and acking. since it was only one bool,
acking would overwrite local changes which would be lost. now
there are two floats, which makes for more junk in the ack object,
but again more robust.

Confidence in my ability to write bug-free code diminishes, but
confidence in my test harness is increasing. I think it's time
to just have two tables in the world, since it makes the test
harness easier to set up mirroring between the two.

I think next is to tackle packet loss in general and see if I
can work in packet-level acks and resubmits into the Bus, and
possibly adaptive waiting. With only two tables, the ownership
change code is pretty much entirely untested anyway, so just need
to find the right lengths of time locally to ensure our packets
get to the other players.