Remove lamport clocks, not really necessary
AFter a bunch of puzzling and halfway through replacing the clocks with vector
clocks, I realized that the isKinematic ownership system should make the
ownership sets of each player disjoint. And as long as they're disjoint,
nothing concurrent is really happening, so no clocks are needed. Just take
the packets at face value, except for the case where the player is locally
holding a tile. That means two players each holding a tile locally will
diverge, with observers switching between whoever receives the packet first.
Once one player drops it though, they'll keep broadcasting and state will
converge again.

Was a cool road down distributed systems theory though. I think the actual
problem that made this same "trust the packet" naive syncing seem broken
and in need of clocks was the local race of locallyMovedTiles and acks.
That seems in a much better place now (with clocks), so I think it's good
to go.